refactor: rename Positions to Position
